Move simple processor tests to simple package
