+++ rs_mcai_worker_sdk/tests/processor/simple/processor.rs
