+++ rs_mcai_worker_sdk/tests/processor/simple/rabbitmq_status.rs
