+++ rs_mcai_worker_sdk/tests/processor/simple/rabbitmq_stop_job.rs
