+++ rs_mcai_worker_sdk/tests/processor/simple/stop_job.rs
