+++ rs_mcai_worker_sdk/tests/tests.rs
@@ -10,11 +10,16 @@ mod amqp {
 #[cfg(not(feature = "media"))]
 mod processor {
   use super::amqp::connection::*;
-  mod rabbitmq_simple_stop_job;
 
-  mod simple_job_processor;
-  mod simple_processor;
-  mod simple_stop_job;
+  mod simple {
+    use super::*;
+
+    mod rabbitmq_stop_job;
+
+    mod job_processor;
+    mod processor;
+    mod stop_job;
+  }
 }
 
 #[cfg(feature = "media")]
